# microservices/product-api-withswagger/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/erankitcs/golang_learning/microservices/product-api-withswagger/handlers
  Product:
    description: Product defines the structure for an API product
    properties:
      description:
        description: the description for this poduct
        maxLength: 10000
        type: string
        x-go-name: Description
      id:
        description: the id for the product
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: the name for this poduct
        maxLength: 255
        type: string
        x-go-name: Name
      price:
        description: the price for the product
        format: float
        minimum: 0.01
        type: number
        x-go-name: Price
      sku:
        description: the SKU for the product
        pattern: '[a-z]+-[a-z]+-[a-z]+'
        type: string
        x-go-name: SKU
    required:
    - name
    - price
    - sku
    type: object
    x-go-package: github.com/erankitcs/golang_learning/microservices/product-api-withswagger/data
host: localhost
info:
  contact:
    email: er.ankit.cs@gmail.com
  description: The purpose of this API is to provide list of products, add new product
    and update the existing product.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: API.
  version: 0.0.1
paths:
  /products:
    get:
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      summary: Return a list of product.
      tags:
      - products
    post:
      operationId: createProduct
      parameters:
      - description: The product to be updated.
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        "200":
          $ref: '#/responses/productsResponse'
        "400":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      summary: Create a new product.
      tags:
      - products
    put:
      description: Update a products details
      operationId: updateProduct
      parameters:
      - description: The product to be updated.
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a product from database
      operationId: deleteProduct
      parameters:
      - description: The id of the product for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
      tags:
      - products
    get:
      operationId: listSingleProduct
      parameters:
      - description: The id of the product for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponse'
      summary: Return a single  product.
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  noContentResponse:
    description: No content is returned by this API endpoint
  productResponse:
    description: Data structure representing a single product
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: A list of product return into the response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"